Fix namespace collisions for headers, add folders to IDEs
One of the source files is math.h, this collides with the cstdlib math.h.  So if math.h was in the include path (which it must be if we are using an IDE) then there is a compile error.

Also added some source files to CMakeLists.txt which I missed before, and added them as folders in IDEs